core: multiple definitions of 'common' data compiling with gcc-10
Winter is coming.  So is gcc-10.

gcc-10 changes the way declared data is defined. Previously it was
defined as common data, now it is BSS.  Common data was silently
merged by the linker.

See https://review.gluster.org/#/c/glusterfs/+/23956/ for a specific
example. The variables fixed in this patch are similar.

There are many changes coming in gcc-10 described in
https://gcc.gnu.org/gcc-10/changes.html

See nm(1) and ld(1) for a description C (common) and B (BSS) and how
they are treated by the linker.

Note: there is still a small chance that gcc-10 will land in Fedora-32,
despite 31 Dec. 2019 having been the deadline for that to happen.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: I212a9c82c4d79012f25b9a4803776ab7a5a8af6d
(cherry picked from commit fc62d0dcc1b7e2f5c45fecacc3dddfa3ec600585)