Fix use of dupreqs with DUPREQ_DELETED state
DUPREQ_DELETED is set just before deleting the dupreq in
nfs_dupreq_delete(). This is called on requests that get dropped! So
these dupreqs should not be used but nfs_dupreq_start() ends up using it
as its state is not DUPREQ_START.

No need to set DUPREQ_DELETED as the dupreq is getting deleted anyway.

Change-Id: Iecd82fae45400d30cea31b4afb1995a0c2d795c0
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>