FSAL_CEPH - Always use the large handle size
Since we check the chk_fscid field to see if it's needed, it has to be
in the wire handle to check.  Always include it in the handle; it will
just be zero if it's unused.

Change-Id: If83e0479413b7a640c6a66ec85ae40296b3f1c72
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit 18928be0831975d63adae764ad4b6e135c4c2e0a)