MDCACHE - Close unexport race for used entries
When we're unexporting, we need to make an entry unreachable if no
exports are referencing it.  However, the refcount is not a suitable
method of determining this, since an OP can be outstanding on the entry,
keeping a ref on it.  If this was the case, we'd leave the entry
dangling, and a new export could reference it, and it's contents could
be bad.

To avoid this, switch from using the refcount to a locked check to make
sure the entry is still not exported.

Change-Id: If7c21274325c23ba1a15797a3be29cf4330e1203
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit e0664947f014966fe291b08c3fa3a1be448558fa)