FSAL_UP: add new "try_release" upcall
Ceph MDS's can request that the client shrink their caps cache in
response to memory pressure, and libcephfs now has a way to set
callbacks that get called when it wants to drop caps for an inode.
Add a new upcall that FSAL_CEPH can call for this.

The upcall will find the entry and test its refcount. If the only
outstanding reference is the sentinel, then it's otherwise unused
and we can free it.

We can't put the last reference while holding the latch though, so
take an extra ref, release the latch and then put it afterward.

Change-Id: Ib195569c88563235014fe1b7e6c37ea1333ceeea
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 58a9114f6b2c48ffa4a04781bdf1faf3ca914f8d)