rados_urls: when built with rados_urls, don't error if lib not installed
The difference between rados-recov and rados-urls is that rados-recov
is explicitly configured in /etc/ganesha/ganesha.conf. If configured,
only then does ganesha try to load the rados-recov shlib. Then, if the
lib is not installed, it issues an error and exits.

But if built with rados-urls, an attempt is always made to load the
shlib, which might not be installed, if only because rados-urls aren't
actually being used. Prior to this both rados-recov and rados-urls
had similar logic and both exited with an error if the shlib couldn't
be loaded.

Now, with this change, just issue a warning that the rados-urls shlib
couldn't be loaded and continue.

Change-Id: I1f22a3237c0573985cfbb840b59317227a69cabb
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>