MDCACHE: remove call of cih_remove_checked in mdcache_unexport
We do not need to unhash the root entry in mdcache_unexport, it was
unhasded by mdcache_lru_cleanup_try_push in the loop if the object
entry is junction node and not used by other export. It's not used
here.
Another case, if the object entry to be root of multiple exports,
here should not unhash and set entry as invalid when the first export
unmount. Until the last export is released, the entry holds the
sentinel reference for recycling and releasing.

Change-Id: Ie15ea032e78c4476bede191e8edcf1619cda4686
Signed-off-by: sepia-liu <liuwei_coder@163.com>
(cherry picked from commit 0693097d61f33d9c8d2bd49b89d448c4f804dd0d)