FSAL_CEPH: track export id as part of the host handle
Typically, ganesha mixes the export_id in with the filehandle before
presenting it to clients, but then strips that out before converting it
to a host handle.

When a handle is reachable via two or more different exports, most FSALs
just keep a single cache entry for it. For FSAL_CEPH, we need to keep
them separate, as the different exports may use different cephx creds.

Add an export_id field to the host handle, and fill out the field from
the op_ctx when converting it from the wire handle. This leaves the
wire handle unchanged, but allows us to keep different cache entries
when the same inode shows up in different exports.

Change-Id: I6426a367061b880ee61c19586f4f54a3bc55ef0a
Signed-off-by: Jeff Layton <jlayton@redhat.com>