NFS: add an on-stack buffer for wire_to_host conversions
Currently, the callers just overwrite the wire handle with the converted
host handle. Overwriting the wire handle like that is somewhat nasty,
but it's also problematic if the host handle is larger than the
wire handle (as will soon be the case for FSAL_CEPH).

Copy the handle into an on-stack buffer first before calling
wire_to_host. Here we assume that the host handle will never be larger
than the max filehandle size for the protocol.

Change-Id: I71791be0c93d5a866da15e50973d1281a0e1ffa2
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 5d3e202cec061612e17eb433852673b4cadd58bf)