Cleanup export refcounting for exports that are only config objects
Until an export is committed during config, it isn't fully instantiated,
and in fact, on an update, a 2nd copy of the export is manifested that
is only used during the config process. Make sure that on releasing
an export that never was committed that we skip some of the cleanup.

This will also assure that the reference counting is correct.

Change-Id: Ica339fddabe702c344cb69ce472b91421447ffed
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>