Fix a regression in 9p implementation
A regression was introduce in the 9p implementation
between V4-dev.22 and V4-dev.23 . As a result, the 9p
mount point appears as ROFS on the client. This is due
to a bad management of op_ctx variable. This commit fixes
this regression.
This commits removes pconn::export_perms field too.

Change-Id: I5181daa6ee9da420ef62af4708a045220d3c7440
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>