FSAL_CEPH: paper over -ENOTCONN return from close when shutting down
When we're shutting down the server, we'll usually abort the connection
first. The mdcache will then try to clean up entries and issue a ->close
to each, and FSAL_CEPH ends up returning -ENOTCONN in that situation
which causes a lot of log spam.

Fix this by just ignoring -ENOTCONN errors in ceph_close_my_fd when
ganesha is shutting down.

Change-Id: If8231998a61e759be4d044f102ae6dbcebfdc975
Reported-and-tested-by: liuwei <liuwei_coder@163.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>