Add missing mutex initialization
The mutex struct is 0-initialized implicitly by the call to gsh_calloc
earlier in this function. This is good enough to make it a functional
mutex by some pthread implementations, but this is not sufficient on at
least macOS. Add an explicit initialization.

Change-Id: I345b1640820fbc591e285dcf375a1832e74270ff
Signed-off-by: Matthew DeVore <matvore@google.com>
(cherry picked from commit 0103714db27cd09aa565f50be3814166fc2b1f87)