ProxyV3: Issue last-chance LOOKUP in READDIRPLUS.
READDIR responses can also be "here's the name" with no handle
or attributes. For example, knfsd (and NetApp) return just the
name for nested mounts/exports. The NFS client is then responsible
for issuing a LOOKUP in response.

Until we add whatever support is needed in Ganesha to return these
sort of partial responses ("just a name", or "name and handle, but
not attributes"), the proxy needs to act more like a full client.

This also finally pushed me to move the readdir entry loop into a
little helper function. The indentation was out of control :).

Change-Id: I96ba4b8791fd25e8273b46df873c812741c2acbe
Signed-off-by: Solomon Boulos <boulos@google.com>