Gather unix uid and gid using wbcGetpwsid
Currently, the primary group is determined from the second sid in the
sids list of the user info. When the flag unix_primary_group is used
in the winbind configuration, the primary gid is instead being looked
up from attribute primaryGroupID while the sids list keeps the same
order. The function wbcGetpwsid asks winbind to determine the primary
gid. It also returns the uid, so that the call to wbcSidToUid can be
omitted.

Change-Id: I919dfb1b32b4f51353efee5a3482cf9aac498b3a
Signed-off-by: Markus Blank-Burian <blankburian@wwu.de>