secinfo: Grab lock only after confirming directory
We need to check the object type before grabbing one of the locks in the
state_hdl. As a bonus, this change deletes an "extra" unlock call needed
as a result of the if block usage.

Fixes crash that occurs on macOS when running st_create.testBlock from
pynfs, because it prevents access of the lock in the wrong side of the
lock union in state_hdl.

Change-Id: I5b40b8fa2ff11076b2858972ab4756458f36c816
Signed-off-by: Matthew DeVore <matvore@google.com>