readdir: Lock only after confirming dir and parms
Before this patch, nfs4_readdir_callback grabbed the junction lock
before checking if the object was a directory or not. Now
nfs4_readdir_callback only locks after checking the object.

Fixes crash that occurs on macOS when running st_readdir.testFirst from
pynfs, because it prevents access of the lock in the wrong side of the
lock union in state_hdl.

Change-Id: I0e2a87c1d81e458c1da8601afef190bed896d78f
Signed-off-by: Matthew DeVore <matvore@google.com>