Readdir - Don't back up dirent for whence-is-name
Early on in whence-is-name, we started caching the last 2 dirents to
make it easier to continue a readdir.  However, we later switched to
caching them all, on an LRU.  This chunk of code was added in the first
set, but not removed in the second.  This will result in a double-unref
of an entry when the chunk is released, since we're starting with an
entry that was already unref'd.

Change-Id: I14995bb9432b60e5d5c35ab4faaa344247529352
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>