FSAL_RGW: allow readdir() to continue past damaged objects
The current FSAL_RGW readdir loop performs a lookup on each
object returned from enumeration, as expected, but stops
enumeration if any lookup fails.

In order to avoid truncated listings in the presence of (e.g.)
a damaged bucket or leaf object, log a warning in the failed
lookup case, but do not terminate the readdir.

Change-Id: Iaf07b6853a3d81ba150d241089f45605ec88920a
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>