Fix op_ctx handling for async FSALs
When we suspend a request, we cannot touch the contents of op_ctx,
because the callback may have already run, and may have freed the
request (and therefore the op_ctx).  All we can do is NULL out op_ctx,
so that future calls don't get a save_op_ctx set wrongly.

Change-Id: I139f68032bfe3233e13b96c8e5bf544258681a68
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>