Clean up log config
Make Default_Log_Level useful. Now if components are not specified, they
will be set to the Default_Log_Level if it's set.

There is also now an original_log_level set from the -N option or if that
option is not set, orginal_log_level is set to EVENT from code default.

If component ALL is specified it overrides all components even if
explicitly set. It ALL is removed, the log levels revert to the
values explictly set, or Debug_Log_Level or original_log_level.

If a component was explicitly set, and then removed, the component will be
reset to Default_Log_Level or original_log_level.

All of this means that a normal environment where the LOG config is empty
can easily be reverted to the original values simply by removing any log
levels that have been configured.

NOTE: A SIGHUP will revert any log levels set by DBUS...

Log level changes are always logged.

RPC_Debug_Level is cleaned up so it is set as now documented.

The logging documentation is improved.

Change-Id: I39648fa626b6307d47151ac171806e2b1aeb0edc
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>