nfs4: encode only valid attrs
I was seeing some warnings like this on Linux clients when accessing
shares that had security labels enabled:

    nfs_setsecurity()  0 security_inode_notifysecctx() -13

The inodes in question had no labels on them, but ganesha was setting
the FATTR4_SEC_LABEL bit in the reply and sending a zero-length label
in the reply. The ceph code will clear ATTR4_SEC_LABEL in that
situation, but the encoder doesn't consult the valid_mask.

Skip encoding any attribute that has a legit attrmask bit, when that
bit is not set in the valid_mask.

Change-Id: I0efb283a0ef0e5fb49bbcb0f1cb03d909161493a
Signed-off-by: Jeff Layton <jlayton@redhat.com>