FSAL_CEPH: override creds when setting sec context on a create
I was seeing problems creating files with the mode 0444 with security
labels enabled. The create would work, but the subsequent setxattr
to set the security label would fail, and that would fail the
operation altogether.

Use the fsal_private field in the op_ctx to pass a flag to setattr2
that tells it to use root creds for the setxattr. Only set that flag
after a create.

Change-Id: I3e9089eb18188f4d36a5f9c5fba48755b9d086cc
Signed-off-by: Jeff Layton <jlayton@redhat.com>