NFS: define a FATTR4_MAX_ATTR_INDEX value
Currently a lot of the nfs_proto_tools.c code is set up to use
FATTR4_XATTR_SUPPORT as the max defined attribute. That's going to be
fragile if we get more new attributes in the future, and it just
doesn't read right.

Define a new FATTR4_MAX_ATTR_INDEX that is currently set to
FATTR4_XATTR_SUPPORT, and use that in the appropriate places in the
code.

Change-Id: I56cc1a79b72f4409acc90543273323f7013c90ce
Signed-off-by: Jeff Layton <jlayton@redhat.com>