Start defining Darwin build based on FreeBSD
Some of the FreeBSD build logic can be used on macOS. So define a
BSDBASED build type and use that where it makes sense. I may have missed
some places where it makes sense to use "#if BSDBASED" rather than "#if
FREEBSD", but these are the obvious ones that are necessary for my macOS
build to minimally function.

Signed-off-by: Matthew DeVore <matvore@google.com>
Change-Id: Idec2f7ac960016c15774490981d2ea21dd2b62dd