Remove librt dependency
Referencing this doesn't work on macOS, and yet it still builds simply
by removing the reference to it in CMakeLists.txt. I was going to make
it conditional, but it actually seems to be completely unused. This will
give you a list of symbols defined by librt:

nm --defined-only /usr/lib/x86_64-linux-gnu/librt.a |
	grep -v '[.:]\|__\|^$' |
	cut -d' ' -f 3 |
	tr '\n' '|' |
	sed 's/|$//' |
	xargs -I % grep --exclude=tags '--exclude=*.rst' --color=always -EIrw % ~/ganesha |
	less -RN

The output of 500+ lines is tedious to go through because it matches
many comments and local variables. But I could not find any function
invocation or pointer reference. So this should be safe to remove.

Change-Id: I4ca17c79ae1d35a8834fc6ece54a62a74eb6bb99
Signed-off-by: Matthew DeVore <matvore@google.com>