CONFIG: Add CONF_ITEM_UI31_ZERO, CONF_ITEM_I64_ZERO, CONF_ITEM_UI64_ZERO
These new config parameter specifiers allow specifying a min/max range
AND allow 0 as a valid value, so the range is (0, min..max).

CONF_ITEM_I64_ZERO is used because parameters that are a time_t are
signed but the range is usually only positive numbers. This now
allows a reasonable minimum to be set while still also allowing 0.

Change-Id: Ib19e0b4eeaabd0db54093f916452c3f0cac031fb
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>