FSAL_GLUSTER: fix double free of caller_garray when nfs3_create fails
The create is performed using tmp_fd until the fsal_obj_handle is
created, at which point, the tmp_fd is copied into globalfd, but if
a subsequent error happened, tmp_fd was closed and then the
fsal_obj_handle was released which closed globalfd. Since the two
shared one caller_garray, double free. The fix is to change my_fd to
refer to globalfd once the copy is accomplished and then ignore tmp_fd.

Change-Id: I4ee6ce428895546e6b683017ee74ac2702c6c681
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>