FSAL_VFS: Make sure fsid is set properly when filling attrs from stat
FSAL_VFS determines the fsid for each file system at the time of
creating the export. The fsid may be the device major/minor but is
more likely to be a UUID, plus is may be modified to be smaller. So
the device major/minor from stat is likely not the correct fsid.

Change-Id: Ife102e1b1719e00c3bcd1c03bb1fc274ca3aef38
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>