Make OPEN create EXCLUSIVE4_1 work correctly
When an exclusive create is done, the attrset bits in the OPEN response
must include the attributes used to store the verifier. Note that for
an EXCLUSIVE4 create, they will be the only bits set. For an EXCLUSIVE4_1
create, the attributes passed in the OPEN will be indicated as well as
the attributes used for the verifier. The client depends on the correect
attrset bits to know that it must issue a SETATTR even if it is otherwise
OK with whatever other attributes are set on create since otherwise the
verifier attributes, atime/mtime in our case, will have an unexpected
value that is derived from the verifier rather than any system clock.

Change-Id: I320f207be8caef5ba97cf8a1ab8c08aea1f534f4
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>