FSAL: Make sure the exclusive create verifier fits
For FSAL_VFS, some filesystems may not allow a full 32 bit atime or mtime.
If this happens and the value is truncated or clamped (set to INT32_MAX),
then try again masking off each 32 bit component to just be 31 bits. We
set a flag in the fsal_filesystem in the future, all verifiers will be
truncated in this way which also assures that the validation of the
verifier works correctly.

Change-Id: I0283fffcde0ef40313eaa657560802a5f77b3549
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>