nsm client always has a caller_name, don't use ssc_client in hash
Clarify the caller_name is always present and also clarify that a deep
copy is done of it.

Since when hashing by IP address, caller_name is the string form of
the IP address, so just hash and search by that and not the actual
IP address.

Because of this, it is also OK if there is no op_ctx->client
since we don't reference it.

NOTE: There WILL always be an op_ctx->client any time we are
establishing a new nsm client since a lock request needs to
come from a client. This allowance is made for SM_NOTIFY
which always comes from loop back and not a real client.

Change-Id: I00ec438e4661dd1fe0c41a7a64e1fca292b80a58
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>