Enable Protocols config in EXPORT_DEFAULTS to work as expected
An EXPORT need not have a Protocols config, in which case the value from
EXPORT_DEFAULTS will be used and if that isn't present, then the value
from NFS_CORE_PARAM will be used.

Also fixup some checking of exports.

Now that an export can be exported NFSv4 by default, we make it NOT an
error to not have a pseudo path, instead we fix up the export to not be
NFSv4.

To make everything work, the Protocols in EXPORT needs to default to
nothing. With this, if Protocols is not specified in CLIENT, EXPORT, or
EXPORT_DEFAULTS, it will defer to the next layer up, ultimately to the
NFS_CORE_PARAM Protocols, which defaults to NFSv3, NFSv4.

Change-Id: If932a602b5dacd266ac2813f84936a97f8ba0f7e
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>