FSAL_LIZARDFS - Convert to new fsal_fd management
NOTE: I have no way of testing this...

This conversion may be incomplete. Unfortunately it must be made so we can
remove the old fsal_fd management code.

For now, this FSAL compiles and is being included, however if maintanence
becomes too much of an issue, it may be deprecated and removed.

Change-Id: I4a0588f2c96c9538eaff9e9d1c2abddf2003941a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>