PROXY_V3: Don't use recursion in READDIR3 and READDIRPLUS3
While this code is in xdr_nfs23.c and looks universal, it's not used
for serving READDIR3 and READDIRPLUS3 because there are special encode
functions that encode each entry into a buffer individually rather than
XDR encoding the entire response. The buffer is also freed as a whole.
So the changed functions would only be used for decode, which would
only apply to FSAL_PROXY_V3.

Change-Id: Ic3d29810f2a5dd4581eca8311057b2f204ba291f
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>