For async I/O the callback must create it's own op_context
We can't pass the orginal caller's op_context, it may still need it.

Just instantiate a simple op_context for the callback.

The callback op_context needs an export reference but we can take
it when we create the op_context because until the protocol callback
is made, we know the protocol request's op_context with its reference
for the export MUST still be valid.

Change-Id: I061c7eed8105b3283a243335b5ea63a86225b85b
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>