Add PSEUDOFS config block
This provides a very simple way to configure the pseudofs root with
several options that are allowed, including specifying a CLIENT list.

This is very useful if a restrictive client list is specified in
EXPORT_DEFAULTS. For example, the following would restrict access to
the pseudofs:

EXPORT_DEFAULTS
{
	Anonymous_uid = -99;
	Anonymous_gid = -99;
	Protocols = 3,4;
	Manage_Gids = false;
	Squash = None;
	Access_Type = RW;
	CLIENT
	{
		Access_Type=RW;
		Clients=192.168.0.0/24,192.168.1.0/24;
	}
	CLIENT
	{
		Access_Type=NONE;
		Clients=*;
	}
}

With this change, the following simple definition restores access to
the pseudofs:

PSEUDOFS
{
	CLIENT
	{
		Clients=*;
	}
}

Thus patch also includes several logging changes that became obvious
while debugging.

Additionally in updating the config documentation for this patch, some
errors were observed and fixed.

Change-Id: I5dccfab272c1ffe2a6750ba2304db706233b9a55
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>