MDCACHE: remove all active entries from LRU L1 and L2
It turns out that entries with elevated refcount clog up L1 and L2
so that entries can not be reaped and the cache grows unbounded.

So outside the LRU reaping process itself, remove any entry with an
elevated refcount from the LRU. This replaces the long term reference
concept.

Additionally, be explicit when the sentinel reference is being released.

This allows for some code cleanup in other areas.

Also, LRU_REQ_INITIAL was supposed to be replaced with LRU_PROMOTE in
all cases.

Also, rename mdcache_dir_entry_t "entry" to "mde_entry" for ease of
auditing the use of the mde_entry.

Also, entries that have only ever been referenced by a dirent will be
returned to L2 when any dirent long term references are released.

Change-Id: If63e31ff2d7490f023b142698e12c444c7432ea8
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>