Failed to get output of option ganesha_stats json
Fixes: https://github.com/nfs-ganesha/nfs-ganesha/issues/1003

This commit fixes the first issue mentioned in the bug.

Issue: The json option of ganesha_stats utility fails and
gives AssertionError. Due to that we cannot use this option

Fix: While converting the data type from dbus to std "int" was
not considered. Due to that AssertionError was raised. Now this
data type is handled.

Change-Id: I35aa6ef23d67e839d4571ba44f7795ec044be3a7
Signed-off-by: Trupti Shete <tshete@ddn.com>