Idmapper: Introduce negative cache for group entries
This change introduces a negative cache for storing group entries
that failed idmapping. The validity and max-count of such entries
is derived from the DIRECTORY_SERVICES section of the Ganesha
config.

This cache is maintained separately from positive cache, to
allow for independent operation, configuration and maintenance.

To avoid code duplication, this CL generalises the existing code
for negative user cache and adapts it for both negative user and
negative group entities.

Change-Id: I2120b730ce3b2bb33d953d657cc9717c8338311b
Signed-off-by: Dipit Grover <dipit@google.com>