build: use standard (i.e. pep517) build and installer
`python setup.py install` (and to a lesser extent `python setup.py build`)
is deprecated, and many build systems are emitting warnings about
support for `setup.py install` being removed.

Note that `python -m installer --prefix ... --destdir ...` doesn't seem
to have any way that I can see to install to /usr/libexec/ganesha/ so
we have to explicitly mv it from /usr/bin to /usr/libexec/... in the
%install stanza of the .spec.

To use outside of an rpmbuild you need python3-wheel, python3-build,
and python3-installer rpms installed on your system. python3-wheel is
available in RHEL and CentOS from the AppStream repo on both C8S and
C9S. python3-build is in EPEL9 but not EPEL8. python3-installer is not
in either EPEL8 or EPEL9 at the time of this writing.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: Ida39a3ee5e6b4b0d3255bfef95601890afd80709