uid2grp: Handle ERANGE return code when calling pw-functions
The pw-functions accept a buffer for holding the strings in
the fetched passwd entry. The buffer's size is derived from
the system config, which should be sufficient in most cases.
To handle the cases where this buffer size is not sufficient,
this CL calls these functions iteratively, while increasing
the buffer size in each iteration.

Change-Id: I5f049684f16d26cb795fc9c42dd23c0593fb5504
Signed-off-by: Dipit Grover <dipit@google.com>