Handle delegation recalls for expired clients
During parallel reads & writes from multiple clients it was seen
that delegation recalls were not progressing as indefinite times
recalls were failing with recall in progress. This was happening
as the client was expired while reserving lease and sd_state was
never getting reset in this process.

This CR handles this issue by setting the state_data.deleg.sd_state
properly and also by wiping off the expired clients in such cases.

Change-Id: Ia94573acdbd83a50fe34afacec0e8c1922ddeb97
Signed-off-by: Rojin George <itsmerojin@gmail.com>