commonlib: Add default flags for generated aces from mode-bits
All should have the READ_ATTR & READ_ACL allowed by default.
Owner in addition Should have the write attr, acl & owner
(actually owner may just change the gid to group it belongs to)

This change is in order to align to posix file-systems behavior (also
the knfsd is doing that).

Signed-off-by: Assaf Yaari <assafy@google.com>
Change-Id: I1c5bab43c9fa87cb073432b604c2a6348614797e