commonlib: Skip redundant aces in mode_to_acl
In fsal_mode_to_acl we generate (or reuse) spedial id aces in order to
represent the owner, group and other mode bits.
Existing (in the middle) such aces, are cleaned from those permissions
bits. If those aces don't include permissions like DELETE or DELETE-CHILD
(which are not covered by the mode bits), they can be skipped and not
being copied.

Signed-off-by: Assaf Yaari <assafy@google.com>
Change-Id: I2fb346a6d2541d5b8692546bb44fbbb4849e95c1