doc: fix cmount_path doc
The line "It must be a subset of the EXPORT { Path } parameter" is really
confusing, everytime i read it, it always implied that it must a subpath of the
path parameter i.e. if the path = /foo/bar then the cmount_path's base|source|root
is /foo/bar and it cannot go up to the hierachy i.e. it cannot be /foo or /.

Signed-off-by: Dhairya Parmar <dparmar@redhat.com>