Release refcount on entry before restarting readdir
When there is a readdir cache hit and a mdcache miss
we create a new mdcache entry and take a refcount on it.
By the time the mdcache entry is created if the readdir
cache entry is invalidated by another thread then we
don't free the ref taken on the created mdcache entry.
We release the reference on the chunks and by
consequence the dirents but that is not enough
because dirents only have a reference to the entry
in the readdir_cb code path.

Change-Id: I07795289084ec6daadf34dfdbbf5aa6769600f89
Signed-off-by: Deepak Arumugam Sankara Subramanian <deepakarumugam.s@nutanix.com>